+++ packages/experimental/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive-experimental@0.12.1...@carbon/ibm-cloud-cognitive-experimental@0.12.2) (2020-12-04)
+
+
+### Bug Fixes
+
+* use correct carbon token for about modal gradient ([#256](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/256)) ([9a9626a](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/9a9626af8d8eb5b5ae26135a755e3047f1e8ea6b))
+
+
+
+
+
+## [0.12.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive-experimental@0.12.0...@carbon/ibm-cloud-cognitive-experimental@0.12.1) (2020-12-03)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive-experimental
+
+
+
+
+
+# [0.12.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive-experimental@0.11.0...@carbon/ibm-cloud-cognitive-experimental@0.12.0) (2020-12-03)
+
+
+### Features
+
+* Breadcrumb with overflow ([#252](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/252)) ([c9acc7a](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/c9acc7af3fa35fa72495e93cb30271ce5124f99f))
+
+
+
+
+
+# [0.11.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive-experimental@0.10.2...@carbon/ibm-cloud-cognitive-experimental@0.11.0) (2020-12-03)
+
+
+### Features
+
+* export modal ([#253](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/253)) ([d7e35ae](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/d7e35ae34567dbfdb99ece3964f558137b687940))
+
+
+
+
+
 ## [0.10.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive-experimental@0.10.1...@carbon/ibm-cloud-cognitive-experimental@0.10.2) (2020-12-02)
 
 **Note:** Version bump only for package @carbon/ibm-cloud-cognitive-experimental